winter1/requirements.txt
Pillow
pyserial
imageio
imageio-ffmpeg
moviepy
numpy
